+++ NEWS.md
@@ -1,3 +1,7 @@
+# ecotraj 0.2.1
+* Update of function 'centerTrajectories' to add flexibility in trajectory centroid definition
+* New vignette to illustrate centering
+
 # ecotraj 0.2.0
 * New function 'trajectorySmoothing' added
 * Update of vignette 'Introduction to to Ecological Trajectory Analysis (ETA)'